+++ libtiff/tif_getimage.c
@@ -3591,13 +3591,13 @@ int TIFFReadRGBATileExt(TIFF *tif, uint32_t col, uint32_t row, uint32_t *raster,
                 read_xsize * sizeof(uint32_t));
         _TIFFmemset(raster + (size_t)(tile_ysize - i_row - 1) * tile_xsize +
                         read_xsize,
-                    0, (tmsize_t)((size_t)sizeof(uint32_t) * (tile_xsize - read_xsize)));
+                    0, (tmsize_t)(sizeof(uint32_t) * (tile_xsize - read_xsize)));
     }
 
     for (i_row = read_ysize; i_row < tile_ysize; i_row++)
     {
         _TIFFmemset(raster + (size_t)(tile_ysize - i_row - 1) * tile_xsize, 0,
-                    (tmsize_t)((size_t)sizeof(uint32_t) * tile_xsize));
+                    (tmsize_t)(sizeof(uint32_t) * tile_xsize));
     }
 
     return (ok);
